+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ScmpFilterContext::{get,set}_badarch_action()` to get/set the default action taken on a syscall for
 an architecture not in the filter.
 - `reset_global_state()` to reset libseccomp's global state.
+- `derive(Hash)` for the most types
 
 ### Changed